Adjust access level for teaching subjects

+++ src/main/java/de/wiss3n/webapp/web/rest/SchoolClassResource.java
@@ -77,11 +77,15 @@ public class SchoolClassResource {
     @Timed
     public ResponseEntity<SchoolClass> updateSchoolClass(@Valid @RequestBody SchoolClass schoolClass) throws URISyntaxException {
         log.debug("REST request to update SchoolClass : {}", schoolClass);
-        this.hasAccess(schoolClass);
 
         if (schoolClass.getId() == null) {
             throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
         }
+
+        if (!schoolClassService.isMySchoolClass(schoolClass.getId())) {
+            throw new BadRequestAlertException("Invalid user", ENTITY_NAME, "not same user");
+        }
+
         SchoolClass result = schoolClassService.save(schoolClass);
         return ResponseEntity.ok()
             .headers(HeaderUtil.createEntityUpdateAlert(ENTITY_NAME, schoolClass.getId().toString()))
@@ -127,15 +131,19 @@ public class SchoolClassResource {
     @Timed
     public ResponseEntity<Void> deleteSchoolClass(@PathVariable Long id) {
         log.debug("REST request to delete SchoolClass : {}", id);
-        schoolClassService.delete(id);
-        return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
+        if (schoolClassService.isMySchoolClass(id)) {
+            schoolClassService.delete(id);
+            return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
     }
 
     /**
      * SEARCH  /_search/school-classes?query=:query : search for the schoolClass corresponding
      * to the query.
      *
-     * @param query the query of the schoolClass search
+     * @param query    the query of the schoolClass search
      * @param pageable the pagination information
      * @return the result of the search
      */
@@ -160,14 +168,7 @@ public class SchoolClassResource {
     public ResponseEntity<List<SchoolClass>> searchActiveSchoolClasses(Pageable pageable) {
         log.debug("REST request to search for a page of SchoolClasses for active");
         Page<SchoolClass> page = schoolClassService.findActive(pageable);
-        HttpHeaders headers = PaginationUtil.generatePaginationHttpHeaders( page, "/api/_search/school-classes/active");
+        HttpHeaders headers = PaginationUtil.generatePaginationHttpHeaders(page, "/api/_search/school-classes/active");
         return new ResponseEntity<>(page.getContent(), headers, HttpStatus.OK);
     }
-
-    private void hasAccess(SchoolClass schoolClass) {
-        if (!schoolClass.getUser().getLogin().equals(SecurityUtils.getCurrentUserLogin().orElse(null))) {
-            throw new BadRequestAlertException("Invalid user", ENTITY_NAME, "not same user");
-        }
-    }
-
 }
